[TWEAK] Seperated EchoBodyHandler from DefaultHandler
